chore: Extract Graphviz DOT logic to module. Merge _extract_refs_from_prop into _extract_prop_relationships. Document safe name characters.
